Added styling for user profiles
Includes changes to both:
	-the small popup user profile
	-the modal popup

Also includes a filter for the tabs in the modal popup for the user profile.
The discord client applies styling to them directly (no classes) which caused me issues.
Simillar behaviour can be mimicked with the selectors such as :active, :hover but im not
sure about the elements initial state.